can roll dice and make a turn again

+++ server/internal/entity/game.go
@@ -5,7 +5,7 @@ type Game struct {
 	GameBoard  *GameBoard `json:"game_board" ` // JSON for gameboard
 	PlayerTurn int        `json:"player_turn"`
 	GameOver   *GameOver  `json:"game_over"`
-	Winner     int        `json:"winner"`
+	Winner     int64      `json:"winner"`
 	Players    []int      `json:"players"` // List of string for the moment, needs changing
 	Full       bool       `json:"full"`
 	LastRoll   []int      `json:"last_roll"` // Last roll of the game, e.g. 5,5
